move emv crypto tests to folder `client/emv/test`
